requirements.txt
asgiref==3.4.1
astroid==2.3.3
certifi==2021.5.30
colorama==0.4.3
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.7.0
djangorestframework==3.12.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.3
pycodestyle==2.5.0
pylint==2.4.4
pytz==2021.1
sqlparse==0.4.2
typed-ast==1.4.0
typing-extensions==3.10.0.2
wincertstore==0.2
wrapt==1.11.2
